+++ jetnn/models/encoders/gnn/__init__.py
@@ -1,6 +1,5 @@
 from .encoder_gcn import GCNEncoder
-from .encoder_gru import GRUEncoder
 from .encoder_ggnn import GGNNEncoder
 from .encoder_rggnn import RGGNNEncoder
-from .encoder_gatv2conv import TransConvEncoder
-from .encoder_myggnn import TypedGGNNEncoder
+from .encoder_transformer_conv import TransformerConvEncoder
+from .encoder_typed_ggnn import TypedGGNNEncoder
