+++ jetnn/models/encoders/gnn/custom_convs/__init__.py
@@ -1,0 +1,1 @@
+from .typed_ggnn import TypedGGNNConv
